refactor: useAuthSWR 위치 변경
