+++ src/components/common/InfoSection.tsx
@@ -1,5 +1,5 @@
 import InfoBanner from "src/components/common/InfoBanner";
-import useAuthSWR from "src/hooks/useAuthSWR";
+import useAuthSWR from "src/hooks/common/useAuthSWR";
 import { CommonListResponse } from "src/types/api/common";
 import { Notice } from "src/types/api/notice";
 import { API_ROUTES } from "src/constants/routes";
