+++ src/pages/[slug]/write.tsx
@@ -8,7 +8,7 @@ import { useEffect, useState } from "react";
 import { PostDetail, PostWritten } from "src/types/api/post";
 import useNotification from "src/contexts/notification/useNotfication";
 import ContentHeaderSection from "src/components/common/ContentHeaderSection";
-import useAuthSWR from "src/hooks/useAuthSWR";
+import useAuthSWR from "src/hooks/common/useAuthSWR";
 import ContentEditor from "src/components/common/ContentEditor";
 import { clientAxios } from "src/utils/common/clientAxios";
 import { AxiosError } from "axios";
